requirements.txt
arrow==0.10.0
binaryornot==0.4.0
chardet==2.3.0
click==6.7
cookiecutter==1.5.1
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.10.2
django-bootstrap3==7.1.0
django-cleanup==0.4.2
django-dia==0.3
django-extensions==1.7.4
django-formtools==1.0
django-sitetree==1.6.0
djangorestframework==3.5.3
future==0.16.0
graphviz==0.5.1
gunicorn==19.7.1
itypes==1.1.0
Jinja2==2.9.5
jinja2-time==0.2.0
MarkupSafe==1.0
paho-mqtt==1.2
poyo==0.4.0
psycopg2==2.6.2
pydotplus==2.0.2
pyparsing==2.1.10
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
whichcraft==0.4.0
